+++ fmi-import/src/test/kotlin/no/mechatronics/sfi/fmi4j/importer/cs/vendors/fmusdk/BouncingBallTest.kt
@@ -1,7 +1,8 @@
-package no.mechatronics.sfi.fmi4j.importer
+package no.mechatronics.sfi.fmi4j.importer.cs.vendors.fmusdk
 
 import no.mechatronics.sfi.fmi4j.TestUtils
 import no.mechatronics.sfi.fmi4j.common.FmiStatus
+import no.mechatronics.sfi.fmi4j.importer.Fmu
 import org.apache.commons.math3.ode.FirstOrderIntegrator
 import org.apache.commons.math3.ode.nonstiff.ClassicalRungeKuttaIntegrator
 import org.apache.commons.math3.ode.nonstiff.EulerIntegrator
@@ -16,20 +17,22 @@ import org.junit.jupiter.api.condition.OS
 import org.slf4j.LoggerFactory
 import java.io.File
 
+
 @EnabledOnOs(OS.WINDOWS)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @EnabledIfEnvironmentVariable(named = "TEST_FMUs", matches = ".*")
-class VanDerPolTest {
+class BouncingBallTest {
 
     private companion object {
-        private val LOG = LoggerFactory.getLogger(VanDerPolTest::class.java)
+
+        private val LOG = LoggerFactory.getLogger(BouncingBallTest::class.java)
     }
 
     private val fmu: Fmu
 
     init {
         val file = File(TestUtils.getTEST_FMUs(),
-                "FMI_2.0/ModelExchange/win64/FMUSDK/2.0.4/vanDerPol/vanDerPol.fmu")
+                "FMI_2.0/ModelExchange/win64/FMUSDK/2.0.4/bouncingBall/bouncingBall.fmu")
         Assertions.assertTrue(file.exists())
         fmu = Fmu.from(file)
     }
@@ -44,22 +47,22 @@ class VanDerPolTest {
         Assertions.assertEquals("2.0", fmu.modelDescription.fmiVersion)
     }
 
-    private fun runFmu(integrator: FirstOrderIntegrator) {
+    private fun runFmu(solver: FirstOrderIntegrator) {
 
-        LOG.info("Using integrator: ${integrator.javaClass.simpleName}")
+        LOG.info("Using solver: ${solver.javaClass.simpleName}")
 
-        fmu.asModelExchangeFmu().newInstance(integrator, loggingOn = true).use { fmu ->
+        fmu.asModelExchangeFmu().newInstance(solver).use { fmu ->
 
-            val x0 = fmu.modelVariables
-                    .getByName("x0").asRealVariable()
+            val h = fmu.modelVariables
+                    .getByName("h").asRealVariable()
 
-            fmu.init(0.0, 0.0)
+            fmu.init()
 
             val macroStep = 1.0 / 10
             while (fmu.currentTime < 1) {
-                val read = x0.read()
+                val read = h.read()
                 Assertions.assertTrue(read.status === FmiStatus.OK)
-                LOG.info("t=${fmu.currentTime}, x0=${read.value}")
+                LOG.info("t=${fmu.currentTime}, h=${read.value}")
                 fmu.doStep(macroStep)
             }
 
@@ -81,5 +84,4 @@ class VanDerPolTest {
     fun testLuther() {
         runFmu(LutherIntegrator(1E-3))
     }
-
 }
